upd iterationsCount data type

+++ src/main/java/edu/project4/utils/RendererRunningConfig.java
@@ -3,7 +3,7 @@ package edu.project4.utils;
 import edu.project4.variationgenerators.Transformation;
 
 public record RendererRunningConfig(
-    int missedIterationsCount, Transformation.Type missedIterationsType,
-    int mainIterationsCount, Transformation.Type mainIterationsType,
+    short missedIterationsCount, Transformation.Type missedIterationsType,
+    short mainIterationsCount, Transformation.Type mainIterationsType,
     int symmetry
 ) {}
